refactor(cli): colocate remaining command tests with subcommands
Migrate tests from commands/__tests__/ to their respective command
directories following the established colocation pattern:
- compose.test.ts → compose/__tests__/index.test.ts
- cook.test.ts → cook/__tests__/index.test.ts
- init.test.ts → init/__tests__/index.test.ts
- onboard.test.ts → onboard/__tests__/index.test.ts
- run.test.ts → run/__tests__/index.test.ts
- setup-claude.test.ts → setup-claude/__tests__/index.test.ts
- usage.test.ts → usage/__tests__/index.test.ts

Update import paths for colocated test structure:
- ../../mocks/server → ../../../mocks/server
- ../command → ../index
- ../../lib/domain/cook-state → ../../../lib/domain/cook-state

All 131 tests continue to pass.
