+++ turbo/apps/cli/src/commands/usage/__tests__/index.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { http, HttpResponse } from "msw";
-import { server } from "../../mocks/server";
-import { usageCommand } from "../usage";
+import { server } from "../../../mocks/server";
+import { usageCommand } from "../index";
 
 describe("usage command", () => {
   const mockExit = vi.spyOn(process, "exit").mockImplementation((() => {
